Fixed resizig bug Added rg250 support

+++ make_release_packages.sh
@@ -1,12 +1,13 @@
 #!/bin/sh
 PROGRAM="meteoroid3d"
-VERSION="1.0.3.1"
+VERSION="1.0.3.2"
 DEST=./build/*
 echo "<html>" > index.htm
 echo "<head>" >> index.htm
 echo "</head>" >> index.htm
 echo "<body>" >> index.htm
 TIME=`date -u +"%d.%m.%Y %R"`
+echo "Version $VERSION" >> index.htm
 echo "Updated at the $TIME." >> index.htm
 echo "<h1>$PROGRAM download links:</h1>" >> index.htm
 echo "<?php" > symlink.php
@@ -39,11 +40,17 @@ do
 					mv "$PROGRAM.opk" ../..
 					echo "<a href=$PROGRAM.opk type=\"application/x-opk+squashfs\">$NAME</a></br>" >> ../../index.htm
 				else
-					zip -r "$PROGRAM-$NAME-$VERSION.zip" * > /dev/null
-					mv "$PROGRAM-$NAME-$VERSION.zip" ../..
-					echo "<a href=$PROGRAM-$NAME-$VERSION.zip>$NAME</a></br>" >> ../../index.htm
-					echo "unlink('$PROGRAM-$NAME.zip');" >> ../../symlink.php
-					echo "symlink('$PROGRAM-$NAME-$VERSION.zip', '$PROGRAM-$NAME.zip');" >> ../../symlink.php
+					if [ $NAME = "rg350" ]; then
+						mksquashfs * "$PROGRAM-$NAME.opk" -all-root -noappend -no-exports -no-xattrs
+						mv "$PROGRAM-$NAME.opk" ../..
+						echo "<a href=$PROGRAM-$NAME.opk type=\"application/x-opk+squashfs\">$NAME</a></br>" >> ../../index.htm
+					else
+						zip -r "$PROGRAM-$NAME-$VERSION.zip" * > /dev/null
+						mv "$PROGRAM-$NAME-$VERSION.zip" ../..
+						echo "<a href=$PROGRAM-$NAME-$VERSION.zip>$NAME</a></br>" >> ../../index.htm
+						echo "unlink('$PROGRAM-$NAME.zip');" >> ../../symlink.php
+						echo "symlink('$PROGRAM-$NAME-$VERSION.zip', '$PROGRAM-$NAME.zip');" >> ../../symlink.php
+					fi
 				fi
 			fi
 		fi
